requirements.txt
asgiref
certifi
crispy-bootstrap5
crispy-tailwind
distlib
dj-database-url
Django
django-allauth
django-cors-headers
django-crispy-forms
django-environ
django-filter
django-tailwind
djangorestframework
djangorestframework-simplejwt
filelock
Markdown
MouseInfo
pillow
pipenv
platformdirs
psycopg2
PyAutoGUI
PyGetWindow
PyJWT
PyMsgBox
pyperclip
PyRect
PyScreeze
pytweening
setuptools
sqlparse
typing_extensions
tzdata
virtualenv
